pyproject.toml
[project]
name = "hisho"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Kazuki Tobita", email = "kazukiti201@gmail.com" }
]
dependencies = [
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
bot = { cmd = "python src/bot.py" }

[tool.hatch.metadata]
allow-direct-references = true
